% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utilities.R
\name{av_seep}
\alias{av_seep}
\title{Calculate the average seepage intensity between radius r1 and r2.}
\usage{
av_seep(r1, r2, y)
}
\arguments{
\item{r1}{Smallest radius (m) (numeric).}

\item{r2}{Largest radius (m) (numeric).}

\item{y}{List with parameters used for the calculations of the groundwater heads and fluxes.}
}
\value{
Average seepage intensity (m/d) (numeric).
}
\description{
Calculate the average seepage intensity between radius r1 and r2.
}
\details{
Positive = downwards flow; negative = upwards flow.
}
\examples{
kD <- c(1000, 1000, 1000)
c <- c(1000, 2000, 3000)
r <- c(1000, 2000)
h <- c(10, 9, 8)
y <- rd_init(kD, c, r, h)
r1 <- 500
r2 <- 1500
av_seep(r1, r2, y)
}
