% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utilities.R
\name{.create_A_matrix}
\alias{.create_A_matrix}
\title{Internal function to create the system matrix.}
\usage{
.create_A_matrix(m, l, kDl, rl)
}
\arguments{
\item{m}{The number of areas for which kD- and c-values are specified (-) (integer).}

\item{l}{Leakage factor (m) (numeric vector).}

\item{kDl}{Quotient of the transmissivity and the leakage factor (m/d) (numeric vector).}

\item{rl}{Matrix with the quotients of the radius of the areas and the leakage factors (-) (matrix).}
}
\value{
System matrix (matrix).
}
\description{
Internal function to create the system matrix.
}
\details{
The dimension of the matrix is (2m, 2*m).
}
