% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utilities.R
\name{.create_b_vector}
\alias{.create_b_vector}
\title{Internal function to create the right hand vector b in the linear equations.}
\usage{
.create_b_vector(h)
}
\arguments{
\item{h}{Polder level in the different areas (m) (numeric vector).}
}
\value{
right hand vector b in the linear equations.
}
\description{
Internal function to create the right hand vector b in the linear equations.
}
\details{
The length of the vector h should be equal to the length of the vectors kD and c.
}
