% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utilities.R
\name{.create_kDl}
\alias{.create_kDl}
\title{Internal function to calculate the quotient of the transmissivity and the leakage factor.}
\usage{
.create_kDl(kD, l)
}
\arguments{
\item{kD}{Transmissivity (m2/d) (numeric vector).}

\item{l}{Leakage factor (m) (numeric vector).}
}
\value{
Quotient of the transmissivity and the leakage factor (m/d) (numeric vector).
}
\description{
Internal function to calculate the quotient of the transmissivity and the leakage factor.
}
\details{
The length of both input vectors should be the same.
}
