% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utilities.R
\name{.create_l}
\alias{.create_l}
\title{Internal function to calculate the leakage factor.}
\usage{
.create_l(kD, c)
}
\arguments{
\item{kD}{Transmissivity (m2/d) (numeric vector).}

\item{c}{Hydraulic resistance of top layer (d) (numeric vector).}
}
\value{
Leakage factor (m) (numeric vector).
}
\description{
Internal function to calculate the leakage factor.
}
\details{
The length of both input vectors should be the same.
}
