% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utilities.R
\name{.create_m}
\alias{.create_m}
\title{Internal function to determine the number of areas for which kD- and c-values are specified.}
\usage{
.create_m(kD)
}
\arguments{
\item{kD}{Transmissivity (m2/d) (numeric vector).}
}
\value{
The number of areas for which kD- and c-values are specified (-) (integer).
}
\description{
Internal function to determine the number of areas for which kD- and c-values are specified.
}
\details{
The number of circular areas is equal to the length of the vector kD minus 1.
}
