% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utilities.R
\name{.create_rl}
\alias{.create_rl}
\title{Internal function to calculate the matrix with the quotients of the radius of the areas and the leakage factors.}
\usage{
.create_rl(r, l)
}
\arguments{
\item{r}{Radius of areas (m) (numeric vector).}

\item{l}{Leakage factor (m) (numeric vector).}
}
\value{
Matrix with the quotients of the radius of the areas and the leakage factors (-) (matrix).
}
\description{
Internal function to calculate the matrix with the quotients of the radius of the areas and the leakage factors.
}
\details{
The dimension of the matrix is (2, length(l)-1).
}
