% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utilities.R
\name{.create_rows}
\alias{.create_rows}
\title{Internal function to create two rows of the system matrix corresponding with area boundary j.}
\usage{
.create_rows(j, m, rl, kDl)
}
\arguments{
\item{j}{area boundary index (-) (integer).}

\item{m}{The number of areas for which kD- and c-values are specified (-) (integer).}

\item{rl}{Matrix with the quotients of the radius of the areas and the leakage factors (-) (matrix).}

\item{kDl}{Quotient of the transmissivity and the leakage factor (m/d) (numeric vector).}
}
\value{
Two rows of the system matrix corresponding with area boundary j (matrix).
}
\description{
Internal function to create two rows of the system matrix corresponding with area boundary j.
}
\details{
The dimension of the matrix is (2, 2*m).
}
