% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utilities.R
\name{.get_coeff}
\alias{.get_coeff}
\title{Internal function to create a list of the parameters corresponding to a distance x to be used for the calculations of the groundwater heads and fluxes (list).}
\usage{
.get_coeff(x, y)
}
\arguments{
\item{x}{Radius (m) (numeric).}

\item{y}{List with parameters used for the calculations of the groundwater heads and fluxes.}
}
\value{
List of the parameters corresponding to a radius x to be used for the calculations of the groundwater heads and fluxes.
}
\description{
Internal function to create a list of the parameters corresponding to a distance x to be used for the calculations of the groundwater heads and fluxes (list).
}
