% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utilities.R
\name{.get_index}
\alias{.get_index}
\title{Internal function to determine the area number corresponding to a radius x.}
\usage{
.get_index(x, r)
}
\arguments{
\item{x}{Radius (m) (numeric).}

\item{r}{Radius of areas (m) (numeric vector).}
}
\value{
Area number corresponding to a radius x (-) (integer).
}
\description{
Internal function to determine the area number corresponding to a radius x.
}
