% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utilities.R
\name{.is_on_border}
\alias{.is_on_border}
\title{Internal function to determine if the radius x corresponds with a area boundary.}
\usage{
.is_on_border(x, r)
}
\arguments{
\item{x}{Radius (m) (numeric).}

\item{r}{Radius of areas (m) (numeric vector).}
}
\value{
TRUE if radius x corresponds with a area boundary (Boolean).
}
\description{
Internal function to determine if the radius x corresponds with a area boundary.
}
