% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utilities.R
\name{.rd_seep}
\alias{.rd_seep}
\title{Calculate the seepage intensity at radius x.}
\usage{
.rd_seep(x, y)
}
\arguments{
\item{x}{Radius (m) (numeric).}

\item{y}{List with parameters used for the calculations of the groundwater heads and fluxes.}
}
\value{
Seepage intensity (m/d) (numeric).
}
\description{
Calculate the seepage intensity at radius x.
}
\details{
Positive = downwards flow; negative = upwards flow.
}
\examples{
kD <- c(1000, 1000, 1000)
c <- c(1000, 2000, 3000)
r <- c(1000, 2000)
h <- c(10, 9, 8)
y <- rd_init(kD, c, r, h)
x <- 500
.rd_seep(x, y)
}
