% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utilities.R
\name{.zero_rows}
\alias{.zero_rows}
\title{Internal function to create a zero two-row matrix.}
\usage{
.zero_rows(m)
}
\arguments{
\item{m}{The number of areas for which kD- and c-values are specified (-) (integer).}
}
\value{
A zero two-row matrix. (matrix).
}
\description{
Internal function to create a zero two-row matrix.
}
\details{
The dimension of the matrix is (2, length(l)-1).
}
