% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utilities.R
\name{rd_init}
\alias{rd_init}
\title{Initialise a list (y) with parameters used for the calculations of the groundwater heads and fluxes.}
\usage{
rd_init(kD, c, r, h)
}
\arguments{
\item{kD}{Transmissivity (m2/d) (numeric vector).}

\item{c}{Hydraulic resistance of top layer (d) (numeric vector).}

\item{r}{Radius of areas (m) (numeric vector).}

\item{h}{Polder level in the different areas (m) (numeric vector).}
}
\value{
List with parameters used for the calculations of the groundwater heads and fluxes.
}
\description{
Initialise a list (y) with parameters used for the calculations of the groundwater heads and fluxes.
}
\examples{
kD <- c(1000, 1000, 1000)
c <- c(1000, 2000, 3000)
r <- c(1000, 2000)
h <- c(10, 9, 8)
y <- rd_init(kD, c, r, h)
}
