% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utilities.R
\name{rd_q}
\alias{rd_q}
\title{Calculate the lateral discharge at radius x (numeric vector).}
\usage{
rd_q(x, y)
}
\arguments{
\item{x}{Radius (m) (numeric vector).}

\item{y}{List with parameters used for the calculations of the groundwater heads and fluxes.}
}
\value{
Lateral discharge (m3/d) (numeric).
}
\description{
Calculate the lateral discharge at radius x (numeric vector).
}
\examples{
kD <- c(1000, 1000, 1000)
c <- c(1000, 2000, 3000)
r <- c(1000, 2000)
h <- c(10, 9, 8)
y <- rd_init(kD, c, r, h)
x <- seq(0,3000,by=100)
rd_q(x, y)
}
